move key spec extraction to clusterProvider
